changes at Siemens

+++ Reports/D4.2/D4.2.2-VV-Implementation/Abs/abs1a.c
@@ -7,7 +7,7 @@
     ensures 0 <= x ==> \result ==  x;
     ensures 0 >  x ==> \result == -x;
 */
-int abs_int(int x)
+short int abs_int(int x)
 {
   return (x >= 0) ? x : -x;
 }
